+++ src/api/withdraw.ts
@@ -1,15 +1,26 @@
 import { AxiosInstance } from 'axios';
-import { Response, WithdrawalParams } from '../types';
+import { Response, SignTransferParams, WithdrawalStatusParams, WithdrawalStatusResponse } from '../types';
 import { parseParams } from '../utils';
 import { WithdrawalResponse } from '../types';
 import { getTransferParams } from './transfer';
 
 export const withdrawalFromL2 = async (
   request: AxiosInstance,
-  data: WithdrawalParams
+  data: SignTransferParams
 ) => {
   const params = await getTransferParams(request, data);
   return request.post<Response<WithdrawalResponse>>('/v1/withdrawalto', {
     ...parseParams(params),
+  });
+};
+
+export const withdrawalStatus = async (
+  request: AxiosInstance,
+  params: WithdrawalStatusParams
+) => {
+  return request.get<Response<WithdrawalStatusResponse[]>>('/v1/withdrawal/status', {
+    params: {
+      ...parseParams(params),
+    },
   });
 };