+++ src/core.ts
@@ -12,6 +12,7 @@ import {
   withdrawalFromL2,
   getBalances,
   order,
+  withdrawalStatus,
 } from './api';
 import {
   erc20Approve,
@@ -21,6 +22,7 @@ import {
 } from './contract';
 import {
   DepositParams,
+  DepositERC20Params,
   Deposit721Params,
   StarkKeyParams,
   SignTransferParams,
@@ -28,7 +30,6 @@ import {
   ApproveErc20Params,
   ErcCommonParams,
   Asset,
-  WithdrawalParams,
   ApproveErc721Params,
   WithdrawalFromL1Params,
   BalanceParams,
@@ -36,6 +37,7 @@ import {
   BalancesParams,
   OrderParams,
   OrderRequestParams,
+  WithdrawalStatusParams,
 } from './types';
 import {
   Env,
@@ -75,24 +77,42 @@ class ReddioCore {
     getVaultID: (args: VaultParams) => {
       return getVaultID(this.request, args);
     },
-    withdrawalFromL2: (args: WithdrawalParams) => {
+    withdrawalFromL2: (args: SignTransferParams) => {
       return withdrawalFromL2(this.request, args);
     },
     withdrawalFromL1: async (args: WithdrawalFromL1Params) => {
       await this.getContractAddress();
       return withdrawalFromL1(this.provider, this.contractAddress!, args);
     },
-    depositERC20: async (args: DepositParams) => {
+    withdrawalStatus: async (args: WithdrawalStatusParams) => {
+      return withdrawalStatus(this.request, args);
+    },
+    depositERC20: async (args: DepositERC20Params) => {
       await this.getContractAddress();
-      return depositERC20(this.provider, this.contractAddress!, args);
+      return depositERC20(
+        this.request,
+        this.provider,
+        this.contractAddress!,
+        args
+      );
     },
     depositETH: async (args: DepositParams) => {
       await this.getContractAddress();
-      return depositETH(this.provider, this.contractAddress!, args);
+      return depositETH(
+        this.request,
+        this.provider,
+        this.contractAddress!,
+        args
+      );
     },
     depositERC721: async (args: Deposit721Params) => {
       await this.getContractAddress();
-      return depositERC721(this.provider, this.contractAddress!, args);
+      return depositERC721(
+        this.request,
+        this.provider,
+        this.contractAddress!,
+        args
+      );
     },
     getBalance: async (args: BalanceParams) => {
       return getBalance(this.request, args);