+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 0.3.0(2022-09-21)
+
+## Features
+
+- add withdrawal status api
+
+## Fix
+
+- balances api params contractAddress should be optional
+- call contract deposit fn should await
+
+## Break change
+
+- Simplify the use of the depositETH api
+- Simplify the use of the depositERC20 api
+- Simplify the use of the depositERC721 api
+- Simplify the use of the transfer and withdrawal api
+
 # 0.2.4(2022-09-14)
 
 ## Features